Agregando por departamento y categoria
