+++ src/components/elements/Header.jsx
@@ -7,7 +7,11 @@ function Header({menuItems, authAction}) {
     const navigate = useNavigate();
     return (
         <header className={'bg-white h-16 flex items-center p-5'}>
-            <h1 className={'flex-none text-3xl text-theme-primary font-bold hover:cursor-pointer'}>{'Fundaciones'}</h1>
+            <h1 className={'flex-none text-3xl text-theme-primary font-bold hover:cursor-pointer'}
+                onClick={()=>{
+                    navigate('/')
+                }}
+            >{'Fundaciones'}</h1>
             <div className="grow">
                 <Menu items={menuItems}/>
             </div>
