+++ src/pages/client/Home/Foundations.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { useEffect , useState} from 'react'
 import { useNavigate } from 'react-router-dom'
+import ListFoundations from '../../../components/elements/ListFoundations'
 import { getAcceptedFoundations } from '../../../services/foundation'
 import { fundaciones } from '../../../utils/mocks/fundacion'
 
@@ -11,9 +12,6 @@ const Fundaciones = () => {
 
     const navigate = useNavigate();
 
-    const handleClick = (id_fundacion) => {
-        navigate(`/fundacion/${id_fundacion}`)
-    }
 
     useEffect(()=>{
         getFoundations();
@@ -22,20 +20,11 @@ const Fundaciones = () => {
     const getFoundations = async()=>{
           const res=await getAcceptedFoundations()
         setFoundations(res);
-        }
+        
+     }
 
     return (
-        foundations.map( fundacion =>(
-            <div key={fundacion.id} onClick={()=>handleClick(fundacion.id)} className="card w-full bg-base-100 shadow-xl hover:cursor-pointer hover:translate-y-1 hover:bg-slate-400">
-                <figure><img src={fundacion.urlImagen} alt="Shoes" /></figure>
-                <div className="card-body">
-                    <h2 className="card-title">{fundacion.nombre}</h2>
-                    <p>{fundacion.descripcion}</p>
-                   
-                </div>
-            </div>
-        )
-        )
+        <ListFoundations  foundations={foundations}/>
     )
 }
